Edited views, added error messages.

+++ AkvelonTask/Controllers/ProjectsController.cs
@@ -20,9 +20,27 @@ namespace AkvelonTask.Controllers
         }
 
         // GET: Projects
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(string sortOrder)
         {
-            return View(await _context.Projects.ToListAsync());
+            ViewData["NameSortParm"] = String.IsNullOrEmpty(sortOrder) ? "name_desc" : "";
+            ViewData["DateSortParm"] = sortOrder == "start_date" ? "start_date" : "end_date";
+            var projects = from p in _context.Projects select p;
+            switch (sortOrder)
+            {
+                case "name_desc":
+                    projects = projects.OrderByDescending(p => p.Name);
+                    break;
+                case "start_date":
+                    projects = projects.OrderByDescending(p => p.StartDate);
+                    break;
+                case "end_date":
+                    projects = projects.OrderByDescending(p => p.EndDate);
+                    break;
+                default:
+                    projects = projects.OrderBy(p => p.Name);
+                    break;
+            }
+            return View(await projects.AsNoTracking().ToListAsync());
         }
 
         // GET: Projects/Details/5
@@ -35,6 +53,7 @@ namespace AkvelonTask.Controllers
 
             var project = await _context.Projects
                 .Include(t => t.Tasks)
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (project == null)
             {
@@ -53,13 +72,22 @@ namespace AkvelonTask.Controllers
         // POST: Projects/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Name,StartDate,EndDate,Status,Priority")] Project project)
+        public async Task<IActionResult> Create([Bind("Name,StartDate,EndDate,Status,Priority")] Project project)
         {
-            if (ModelState.IsValid)
+            try
             {
-                _context.Add(project);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(project);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+            }
+            catch (DbUpdateException)
+            {
+                ModelState.AddModelError("", "Unable to save changes. " +
+                    "Try again, and if the problem persists " +
+                    "see your system administrator.");
             }
             return View(project);
         }
@@ -96,25 +124,20 @@ namespace AkvelonTask.Controllers
                 {
                     _context.Update(project);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateException)
                 {
-                    if (!ProjectExists(project.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    ModelState.AddModelError("", "Unable to save changes. " +
+                        "Try again, and if the problem persists " +
+                        "see your system administrator.");
                 }
-                return RedirectToAction(nameof(Index));
             }
             return View(project);
         }
 
         // GET: Projects/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -122,10 +145,17 @@ namespace AkvelonTask.Controllers
             }
 
             var project = await _context.Projects
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (project == null)
             {
                 return NotFound();
+            }
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                    "Delete failed. Try again, and if the problem persists " +
+                    "see your system administrator.";
             }
 
             return View(project);
@@ -137,9 +167,22 @@ namespace AkvelonTask.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var project = await _context.Projects.FindAsync(id);
-            _context.Projects.Remove(project);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (project == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            try
+            {
+                _context.Projects.Remove(project);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log.)
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
 
         public IActionResult CreateTask()
@@ -149,13 +192,22 @@ namespace AkvelonTask.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> CreateTask([Bind("Id,Name,Description,Status,Priority")] Models.Task task)
+        public async Task<IActionResult> CreateTask([Bind("Name,Description,Status,Priority")] Models.Task task)
         {
-            if (ModelState.IsValid)
+            try
             {
-                _context.Add(task);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(task);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+            }
+            catch (DbUpdateException)
+            {
+                ModelState.AddModelError("", "Unable to save changes. " +
+                    "Try again, and if the problem persists " +
+                    "see your system administrator.");
             }
             return View(task);
         }
